Moving parse_args into config module
